Fixed nasty bug causing machine-dependent discrepancies.
The bug manifested as the thinkpad succeeding in factoring the number,
while the macbook (and clyde) spat it out as composite. It turned out
that the problem was that the x value of a rel_t was declared as
uint32_t when it should have been int32_t. I was using mpz_set_si to
load these values; mpz_set_si takes something of type 'long int' as its
parameter, which on the thinkpad was equivalent to int32_t. It was
interpreting the uint as an signed int, and was getting the right
result. However, on the other machines, a long int is 64 bits, and
interpreting an unsigned 32 bit number as a signed 64 bit number had
problems. YUCK.